initial implementation of CAS instructions

+++ include/RevMem.h
@@ -50,31 +50,30 @@
 #define _REVMEM_BASE_ 0x00000000
 #endif
 
-#define _STACK_SIZE_ ( size_t{ 1024 * 1024 } )
+#define _STACK_SIZE_         ( size_t{ 1024 * 1024 } )
 
-#define Z_ADDR_MASK  0xFFFFFFFFFF
-#define Z_ZONE_MASK  0b111
-#define Z_PREC_MASK  0x1FFF
-#define Z_SP_MASK    0b01
-#define Z_VIEW_MASK  0b01
-#define Z_SEG_MASK   0x3F
-#define Z_ADDR_SHIFT 0x00
-#define Z_ZONE_SHIFT 40
-#define Z_PREC_SHIFT 43
-#define Z_SP_SHIFT   56
-#define Z_VIEW_SHIFT 57
-#define Z_SEG_SHIFT  58
+#define Z_ADDR_MASK          0xFFFFFFFFFF
+#define Z_ZONE_MASK          0b111
+#define Z_PREC_MASK          0x1FFF
+#define Z_SP_MASK            0b01
+#define Z_VIEW_MASK          0b01
+#define Z_SEG_MASK           0x3F
+#define Z_ADDR_SHIFT         0x00
+#define Z_ZONE_SHIFT         40
+#define Z_PREC_SHIFT         43
+#define Z_SP_SHIFT           56
+#define Z_VIEW_SHIFT         57
+#define Z_SEG_SHIFT          58
 
 // KL: Since depth == NHARTS, and each HART can only be requesting one spawn presently, I'm simplifying.
 //#define SP_REQ_FIFO_D       = NHARTS;
 //#define SP_REQ_FIFO_D_LOG2  = LOG_NHARTS;
-#define SP_TRACK_FIFO_D       32
-#define SP_TRACK_FIFO_D_LOG2  5
-#define RING_REQ_FIFO_D       32
-#define RING_REQ_FIFO_D_LOG2  5
-#define WB_FIFO_D             16
-#define WB_FIFO_D_LOG2        4
-
+#define SP_TRACK_FIFO_D      32
+#define SP_TRACK_FIFO_D_LOG2 5
+#define RING_REQ_FIFO_D      32
+#define RING_REQ_FIFO_D_LOG2 5
+#define WB_FIFO_D            16
+#define WB_FIFO_D_LOG2       4
 
 namespace SST::RevCPU {
 
@@ -201,24 +200,16 @@ public:
   void SetTracer( RevTracer* tracer ) { Tracer = tracer; }
 
   /// RevMem: Allocate and zero thread done vector
-  void ThreadQuitSetup(uint32_t num_harts){
-    threadQuits.assign(num_harts, 0);  
-  }
-  
+  void ThreadQuitSetup( uint32_t num_harts ) { threadQuits.assign( num_harts, 0 ); }
+
   /// RevMem: Mark quitting thread
-  void IssueThreadQuit(uint32_t hart){
-    threadQuits[hart] = true;
-  }
+  void IssueThreadQuit( uint32_t hart ) { threadQuits[hart] = true; }
 
   /// RevMem: Finish quitting -- unmark thread quit at minimum
-  void FinalizeThreadQuit(uint32_t hart){
-    threadQuits[hart] = false;
-  }
+  void FinalizeThreadQuit( uint32_t hart ) { threadQuits[hart] = false; }
 
   /// RevMem: Check for quitting thread
-  bool CheckThreadQuit(uint32_t hart){
-    return threadQuits[hart];
-  }
+  bool CheckThreadQuit( uint32_t hart ) { return threadQuits[hart]; }
 
   // ----------------------------------------------------
   // ---- Base Memory Interfaces
@@ -416,10 +407,10 @@ public:
   // ----------------------------------------------------
   // ---- FORZA Interfaces
   // ----------------------------------------------------
-  
+
   /// FORZA: set the Zone Ring object
-  void setZRing( Forza::RingNetAPI* R ) { zoneRing = R;  }
-  
+  void setZRing( Forza::RingNetAPI* R ) { zoneRing = R; }
+
   /// FORZA: set the ZOP NIC object
   void setZNic( Forza::zopAPI* Z ) { zNic = Z; }
 
@@ -483,19 +474,19 @@ public:
   void DumpThreadMem( const uint64_t bytesPerRow = 16, std::ostream& outputStream = std::cout );
 
   /// FORZA: determines if the current spawn instruction has been previously fired
-  bool ThreadQIsActive(uint32_t Hart);
+  bool ThreadQIsActive( uint32_t Hart );
 
   /// FORZA: determines if the current spawn instruction FSM is complete
-  bool ThreadQIsComplete(uint32_t Hart);
+  bool ThreadQIsComplete( uint32_t Hart );
 
   /// FORZA: injects a new thread into the ThreadQ FSM
-  bool ThreadQInsert(uint32_t Hart, uint64_t TPC, uint64_t X31);
+  bool ThreadQInsert( uint32_t Hart, uint64_t TPC, uint64_t X31 );
 
   /// FORZA: process the ThreadQ FSM: This should ONLY be called from the CPU's clock method
   bool ThreadQProcess();
 
   /// FORZA: receive response from ZEN during spawn FSM
-  void ThreadQReceiveZen( Forza::ringEvent * ev );
+  void ThreadQReceiveZen( Forza::ringEvent* ev );
 
   /// FORZA: send word to Zen for spawning
   void FSMSendZenWord( uint16_t Hart, uint64_t Datum );
@@ -514,7 +505,7 @@ private:
   bool ZOP_READMem( uint32_t Hart, uint64_t Addr, size_t Len, void* Target, const MemReq& req, RevFlag flags );
 
   /// FORZA : send READ request to Zen for HART
-  void FSMReadZen(uint16_t Hart);
+  void FSMReadZen( uint16_t Hart );
 
   /// FORZA: send a WRITE request
   bool ZOP_WRITEMem( uint32_t Hart, uint64_t Addr, size_t Len, const void* Data, RevFlag flags );
@@ -543,10 +534,10 @@ private:
   RevOpts*            opts{};      ///< RevMem: options object
   RevMemCtrl*         ctrl{};      ///< RevMem: memory controller object
   SST::Output*        output{};    ///< RevMem: output handler
- 
-  Forza::RingNetAPI* zoneRing{}; ///< RevMem: FORZA RingNet; Necessary for spawn messages
-  Forza::zopAPI* zNic{};   ///< RevMem: FORZA ZOP NIC
-  bool           isRZA{};  ///< RevMem: FORZA RZA flag; true if this device is an RZA
+
+  Forza::RingNetAPI* zoneRing{};  ///< RevMem: FORZA RingNet; Necessary for spawn messages
+  Forza::zopAPI*     zNic{};      ///< RevMem: FORZA ZOP NIC
+  bool               isRZA{};     ///< RevMem: FORZA RZA flag; true if this device is an RZA
 
   std::vector<std::shared_ptr<MemSegment>> MemSegs{};        // Currently Allocated MemSegs
   std::vector<std::shared_ptr<MemSegment>> FreeMemSegs{};    // MemSegs that have been unallocated
@@ -579,7 +570,7 @@ private:
   uint64_t heapstart{};   ///< RevMem: top of the stack
   uint64_t stacktop{};    ///< RevMem: top of the stack
 
-  std::vector<bool> threadQuits; ///< RevMem: array of threads to enable quit signaling to RevCore  
+  std::vector<bool> threadQuits;  ///< RevMem: array of threads to enable quit signaling to RevCore
 
   std::vector<uint64_t>                                     FutureRes{};  ///< RevMem: future operation reservations
   std::unordered_map<uint32_t, std::pair<uint64_t, size_t>> LRSC{};       ///< RevMem: load reserve/store conditional set
@@ -594,8 +585,6 @@ private:
   bool                                                                PhysAddrLogging{};
   std::string                                                         outputFile;
 
-
-
   // KL: WIP SPAWN/PZOP LOGIC HEREIN
   // FORZA Thread State
   enum class ThreadQState {
@@ -603,18 +592,18 @@ private:
     Complete = 1,
   };
   // Hart, TPC, X31, ThreadQState enum
-  std::vector<std::tuple<uint32_t,uint64_t,uint64_t,ThreadQState>>  ThreadQ;
-  
-  // I'm using this to make it so an entire vector isn't traversed per check-in from 
-  // HART on its spawns. 
-  std::map<uint32_t, ThreadQState> InsertionStatTrack; 
+  std::vector<std::tuple<uint32_t, uint64_t, uint64_t, ThreadQState>> ThreadQ;
+
+  // I'm using this to make it so an entire vector isn't traversed per check-in from
+  // HART on its spawns.
+  std::map<uint32_t, ThreadQState> InsertionStatTrack;
 
   /* // Hold control info while status is read
    typedef struct packed {
       logic        pzp_valid;
       logic        spn_valid;
       tcb_t        tcb;
-      logic [4:0]  func5; // this has something to do with instructions to run -- and apparently does take 3 registers at times. It shows up in instruction formats. 
+      logic [4:0]  func5; // this has something to do with instructions to run -- and apparently does take 3 registers at times. It shows up in instruction formats.
       logic [63:0] rs1_data;
       logic [63:0] rs2_data;
       logic [63:0] rs3_data;
@@ -623,19 +612,19 @@ private:
 
   // Every bool herein is tied to a 'logic' in RTL, which supports 4 states.
   // However, in terms of behavior, they seem to be evaluated in terms of on/off, so I use a bool here instead
-  // of managing 2 bits. 
-  struct spawnTracker{
-    bool pzpValid;
-    bool spnValid;
+  // of managing 2 bits.
+  struct spawnTracker {
+    bool     pzpValid;
+    bool     spnValid;
     uint32_t tcb;
-    uint8_t func5; // Technically too big -- defined as 4 bits.
+    uint8_t  func5;  // Technically too big -- defined as 4 bits.
     uint64_t rs1Data;
     uint64_t rs2Data;
     uint64_t rs3Data;
-    uint8_t rd; // I think this is too big -- refers to a register number, defined in ZAP doc as 4 bits. 
+    uint8_t  rd;  // I think this is too big -- refers to a register number, defined in ZAP doc as 4 bits.
   };
 
-/*
+  /*
    typedef struct  packed {
       logic        valid;
       exec_op_t    op;
@@ -649,67 +638,66 @@ private:
    } spn_pzp_late_t;     //used
 */
 
-  struct spawnPZOPInput{
-    bool valid;
-    uint8_t op;
+  struct spawnPZOPInput {
+    bool     valid;
+    uint8_t  op;
     uint32_t hart;
     uint32_t tcb;
-    uint8_t func5;
+    uint8_t  func5;
     uint64_t rs1Data;
     uint64_t rs2Data;
     uint64_t rs3Data;
-    uint8_t rd;
+    uint8_t  rd;
   };
 
-/* for
+  /* for
    inst_sp_req_q (
                   .clk         (clk),
                   .resetn      (resetn),
                   .i_data_in   (sp_req_q_wdata),
                   .i_write     (sp_req_q_wr),
                   .o_data_out  (sp_req_q_rdata),
-                  // verilator lint_off PINCONNECTEMPTY 
+                  // verilator lint_off PINCONNECTEMPTY
                   .o_full      (),
                   .o_near_full (),
                   .o_used      (),
-                  // lint_on 
+                  // lint_on
                   .o_empty     (sp_req_q_empty),
                   .i_read      (sp_req_q_rd)
                   );
 		  */
   // request queue
-  std::queue<spawnPZOPInput>  sp_req_q; 
+  std::queue<spawnPZOPInput> sp_req_q;
   // data waiting to be used for request fulfillment
   std::queue<spawnTracker> sp_track_q;
 
-
-  struct zenSpawnStatus{
-    uint32_t hart; 
-    bool zenEn;
-    bool SPWriteError;
-    bool busy;
+  struct zenSpawnStatus {
+    uint16_t hart;
+    bool     zenEn;
+    bool     SPWriteError;
+    bool     busy;
   };
-  
+
   std::queue<zenSpawnStatus> ring_rtn_q;
 
   enum class FSMState {
-    IDLE = 0,
-    STATUS_RD = 1,
-    STATUS_RTN = 2,
+    IDLE        = 0,
+    STATUS_RD   = 1,
+    STATUS_RTN  = 2,
     SEND_PC_TCB = 3,
-    SEND_WRD_1 = 4,
-    SEND_WRD_2 = 5,
-    SEND_WRD_3 = 6,
-    SEND_WRD_4 = 7,
-    WR_BACK = 8,
+    SEND_WRD_1  = 4,
+    SEND_WRD_2  = 5,
+    SEND_WRD_3  = 6,
+    SEND_WRD_4  = 7,
+    WR_BACK     = 8,
   };
 
-  FSMState sp_state = FSMState::IDLE;
+  FSMState sp_state      = FSMState::IDLE;
   FSMState next_sp_state = FSMState::IDLE;
 
-  const char* FSMReportState(FSMState state);
+  const char* FSMReportState( FSMState state );
 
-/*   typedef enum [1:0] {
+  /*   typedef enum [1:0] {
                  NA,
                  CSR_READ,
                  CSR_WRITE,
@@ -718,9 +706,9 @@ private:
 
   // use Forza::ringMsgT to represent csr_func_t
   // update == write, read == read, return_data isn't used here but is a simulator-only command
-  // for returning data from ring to service read. 
+  // for returning data from ring to service read.
 
-/*typedef enum logic [ZXB_ZC_WID-1:0]
+  /*typedef enum logic [ZXB_ZC_WID-1:0]
 {
   ZAP0  = 0,
   ZAP1  = 1,
@@ -733,9 +721,9 @@ private:
   ZEN   = 8,
   ZQM   = 9
 } zone_comp_t;*/
-// Satisfied by Forza::zopCompID
+  // Satisfied by Forza::zopCompID
 
-/*   typedef struct packed
+  /*   typedef struct packed
                   {
                      csr_func_t                        func;
                      zone_comp_t                       dev_no;
@@ -747,43 +735,42 @@ private:
                      logic [ZXB_DATA_WID-1:0]          data;
                   } csr_ring_t;*/
 
-
-  // Might be defined elsewhere? 
+  // Might be defined elsewhere?
   struct ringMessage {
-    Forza::ringMsgT func;
+    Forza::ringMsgT  func;
     Forza::zopCompID dev_no;
-    uint32_t aid;
-    uint32_t addr;
-    bool sp_vld;
-    uint32_t zapid;
-    uint32_t hartid;
-    uint32_t data;
+    uint32_t         aid;
+    uint32_t         addr;
+    bool             sp_vld;
+    uint32_t         zapid;
+    uint32_t         hartid;
+    uint32_t         data;
   };
 
   std::queue<ringMessage> ring_req_q;
 
-/*    typedef struct packed { // Send to pc_tcb_mux without tcb update info
+  /*    typedef struct packed { // Send to pc_tcb_mux without tcb update info
       logic        reg_valid;
       hart_t       hart;
       regnum_t     reg_no;
       logic [63:0] data;
    } wb_tcb_pc_t;          //used*/
 
-  struct wb_tcb_pc_t{
-    bool regValid;
+  struct wb_tcb_pc_t {
+    bool     regValid;
     uint32_t hartID;
     uint32_t regNumber;
     uint64_t data;
   };
 
   std::queue<wb_tcb_pc_t> wbQueue;
-  spawnTracker sp_track_busy_data;
-  bool sp_track_data_vld;
-  uint16_t sp_wr_hart;
-  uint16_t status_rtn_hart;
-  bool status_busy;
-  bool spn_valid;
-  bool pzp_valid;
+  spawnTracker            sp_track_busy_data;
+  bool                    sp_track_data_vld;
+  uint16_t                sp_wr_hart;
+  uint16_t                status_rtn_hart;
+  bool                    status_busy;
+  bool                    spn_valid;
+  bool                    pzp_valid;
   // std::ofstream output_file;
   // std::ofstream input_file;
